requirements.txt
Flask==2.3.3
pypdf==3.17.4
Werkzeug==2.3.7
python-docx==1.1.0
gunicorn==21.2.0